+++ apps/core/src/components/pure/trash-button-group/styles.css.ts
